Demos

+++ WroldCupDemo/Models/Team.cs
@@ -9,14 +9,11 @@ namespace WorldCupDemo.Models
     public class Team
     {
         public int Id { get; set; }
-
         public string Name { get; set; }
-
         public string GroupName { get; set; }
         public int Appearances { get; set; }
-        public int Titles { get; set; }
+        public byte Titles { get; set; }
         public int Rank { get; set; }
-
         public int Finals { get; set; }
     }
 }
